feat(express): add support for binding to an IPC socket (unix socket)
Sometimes it can be desirable to further restrict access to the BitGo
Express API endpoint beyond that of a localhost TCP connection. An
effective low-effort way of accomplishing further restriction is by
using a Unix domain socket instead. For example, the Unix domain socket
can be placed in a part of the file hierarchy where only the Unix
users/groups that require access to BitGo Express have access. Other
users/groups will be unable to access the Unix socket.

Ticket: BG-28829